+++ simple-stack/src/main/java/com/zhuinden/simplestack/BackstackDelegate.java
@@ -457,6 +457,20 @@ public class BackstackDelegate {
     }
 
     /**
+     * Attempts to look-up the service in all currently existing scopes, starting from the last added scope.
+     * If the service is not found, an exception is thrown.
+     *
+     * @param serviceTag the tag of the service
+     * @param <T>        the type of the service
+     * @return the service
+     * @throws IllegalStateException if the service doesn't exist in any scope
+     */
+    @NonNull
+    public <T> T lookupService(@NonNull String serviceTag) {
+        return getManager().lookupService(serviceTag);
+    }
+
+    /**
      * Returns the {@link BackstackManager}. If called before {@link BackstackDelegate#onCreate(Bundle, Object, List)}, it throws an exception.
      *
      * @return the backstack manager